feat(doge): add doge consensus and genesis parameters
