+++ bitcoin/src/consensus/params.rs
@@ -75,12 +75,6 @@ pub static TESTNET4: Params = Params::TESTNET4;
 pub static SIGNET: Params = Params::SIGNET;
 /// The regtest parameters.
 pub static REGTEST: Params = Params::REGTEST;
-/// The dogecoin mainnet parameters.
-pub static DOGECOIN: Params = Params::DOGECOIN;
-/// The dogecoin testnet parameters.
-pub static DOGECOIN_TESTNET: Params = Params::DOGECOIN_TESTNET;
-/// The dogecoin regtest parameters.
-pub static DOGECOIN_REGTEST: Params = Params::DOGECOIN_REGTEST;
 
 #[allow(deprecated)] // For `pow_limit`.
 impl Params {
@@ -190,57 +184,6 @@ impl Params {
         no_pow_retargeting: true,
     };
 
-    /// The Dogecoin mainnet parameters.
-    pub const DOGECOIN: Params = Params {
-        network: Network::Dogecoin,
-        bip16_time: 1333238400,                 // Apr 1 2012
-        bip34_height: 1034383, // 80d1364201e5df97e696c03bdd24dc885e8617b9de51e453c10a4f629b1e797a
-        bip65_height: 3464751, // 34cd2cbba4ba366f47e5aa0db5f02c19eba2adf679ceb6653ac003bdc9a0ef1f
-        bip66_height: 1034383, // 80d1364201e5df97e696c03bdd24dc885e8617b9de51e453c10a4f629b1e797a
-        rule_change_activation_threshold: 9576, // 95% of 10,080
-        miner_confirmation_window: 10080, // 60 * 24 * 7 = 10,080 blocks, or one week
-        pow_limit: Target::MAX_ATTAINABLE_MAINNET_DOGE,
-        max_attainable_target: Target::MAX_ATTAINABLE_MAINNET_DOGE,
-        pow_target_spacing: 60,           // 1 minute
-        pow_target_timespan: 4 * 60 * 60, // pre-digishield: 4 hours
-        allow_min_difficulty_blocks: false,
-        no_pow_retargeting: false,
-    };
-
-    /// The Dogecoin testnet parameters.
-    pub const DOGECOIN_TESTNET: Params = Params {
-        network: Network::DogecoinTestnet,
-        bip16_time: 1333238400,                 // Apr 1 2012
-        bip34_height: 708658, // 21b8b97dcdb94caa67c7f8f6dbf22e61e0cfe0e46e1fff3528b22864659e9b38
-        bip65_height: 1854705, // 955bd496d23790aba1ecfacb722b089a6ae7ddabaedf7d8fb0878f48308a71f9
-        bip66_height: 708658, // 21b8b97dcdb94caa67c7f8f6dbf22e61e0cfe0e46e1fff3528b22864659e9b38
-        rule_change_activation_threshold: 2880, // 2 days (note this is significantly lower than Bitcoin standard)
-        miner_confirmation_window: 10080,       // 60 * 24 * 7 = 10,080 blocks, or one week
-        pow_limit: Target::MAX_ATTAINABLE_TESTNET_DOGE,
-        max_attainable_target: Target::MAX_ATTAINABLE_TESTNET_DOGE,
-        pow_target_spacing: 60,           // 1 minute
-        pow_target_timespan: 4 * 60 * 60, // pre-digishield: 4 hours
-        allow_min_difficulty_blocks: true,
-        no_pow_retargeting: false,
-    };
-
-    /// The Dogecoin regtest parameters.
-    pub const DOGECOIN_REGTEST: Params = Params {
-        network: Network::DogecoinRegtest,
-        bip16_time: 1333238400,  // Apr 1 2012
-        bip34_height: 100000000, // not activated on regtest
-        bip65_height: 1351,
-        bip66_height: 1251,                    // used only in rpc tests
-        rule_change_activation_threshold: 540, // 75%
-        miner_confirmation_window: 720,
-        pow_limit: Target::MAX_ATTAINABLE_REGTEST_DOGE,
-        max_attainable_target: Target::MAX_ATTAINABLE_REGTEST_DOGE,
-        pow_target_spacing: 1,            // regtest: 1 second blocks
-        pow_target_timespan: 4 * 60 * 60, // pre-digishield: 4 hours
-        allow_min_difficulty_blocks: true,
-        no_pow_retargeting: true,
-    };
-
     /// Creates parameters set for the given network.
     pub const fn new(network: Network) -> Self {
         match network {
@@ -249,9 +192,6 @@ impl Params {
             Network::Testnet4 => Params::TESTNET4,
             Network::Signet => Params::SIGNET,
             Network::Regtest => Params::REGTEST,
-            Network::Dogecoin => Params::DOGECOIN,
-            Network::DogecoinTestnet => Params::DOGECOIN_TESTNET,
-            Network::DogecoinRegtest => Params::DOGECOIN_REGTEST,
         }
     }
