+++ bitcoin/src/network.rs
@@ -52,9 +52,8 @@ impl From<Network> for NetworkKind {
         use Network::*;
 
         match n {
-            Bitcoin | Dogecoin => NetworkKind::Main,
-            Testnet | Testnet4 | Signet | Regtest | DogecoinTestnet | DogecoinRegtest =>
-                NetworkKind::Test,
+            Bitcoin => NetworkKind::Main,
+            Testnet | Testnet4 | Signet | Regtest => NetworkKind::Test,
         }
     }
 }
@@ -78,12 +77,6 @@ pub enum Network {
     Signet,
     /// Bitcoin's regtest network.
     Regtest,
-    /// Mainnet Dogecoin.
-    Dogecoin,
-    /// Dogecoin's testnet network.
-    DogecoinTestnet,
-    /// Dogecoin's regtest network.
-    DogecoinRegtest,
 }
 
 impl Network {
@@ -131,9 +124,6 @@ impl Network {
             Network::Testnet4 => "testnet4",
             Network::Signet => "signet",
             Network::Regtest => "regtest",
-            Network::Dogecoin => "doge_main",
-            Network::DogecoinTestnet => "doge_test",
-            Network::DogecoinRegtest => "doge_regtest",
         }
     }
 
@@ -195,9 +185,6 @@ impl Network {
             Network::Testnet4 => &Params::TESTNET4,
             Network::Signet => &Params::SIGNET,
             Network::Regtest => &Params::REGTEST,
-            Network::Dogecoin => &Params::MAINNET,
-            Network::DogecoinTestnet => &Params::TESTNET,
-            Network::DogecoinRegtest => &Params::REGTEST,
         }
     }
 
@@ -210,9 +197,6 @@ impl Network {
             Network::Testnet4 => "testnet4",
             Network::Signet => "signet",
             Network::Regtest => "regtest",
-            Network::Dogecoin => "dogecoin",
-            Network::DogecoinTestnet => "dogecoin_testnet",
-            Network::DogecoinRegtest => "dogecoin_regtest",
         }
     }
 }
